Tests/LibWeb/Text/expected/Editing/execCommand-superscript.txt

Div contents: "foobar" state: false selection: #document 0 #document 0
Div contents: "foo<sup>bar</sup>" state: true selection: #text 0 #text 3
Div contents: "<sup>foobar</sup>" state: true selection: #text 0 #text 3
Div contents: "<sup>foo</sup>bar" state: false selection: #text 0 #text 3
